Create module element-web-module-banner
