+++ packages/element-web-module-api/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@element-hq/element-web-module-api",
     "type": "module",
-    "version": "0.1.5",
+    "version": "0.2.0",
     "description": "Module API surface for element-web",
     "repository": {
         "type": "git",
@@ -34,6 +34,7 @@
         "@microsoft/api-extractor": "^7.49.1",
         "@types/node": "^22.10.7",
         "@types/react": "^19",
+        "@types/react-dom": "^19.0.4",
         "@types/semver": "^7.5.8",
         "@vitest/coverage-v8": "^3.0.4",
         "matrix-web-i18n": "^3.3.0",
