+++ packages/element-web-module-api/src/index.ts
@@ -6,6 +6,8 @@ Please see LICENSE files in the repository root for full details.
 */
 
 export { ModuleLoader, ModuleIncompatibleError } from "./loader";
-export type { Api, Module, ModuleFactory, Config, ConfigApi } from "./api";
-export type * from "./legacy-modules";
-export type * from "./legacy-customisations";
+export type { Api, Module, ModuleFactory } from "./api";
+export type { Config, ConfigApi } from "./api/config";
+export type { I18nApi, Variables, Translations } from "./api/i18n";
+export type * from "./api/legacy-modules";
+export type * from "./api/legacy-customisations";
